[DBTablesMonitoring] Delete unused deprecated method: getHostInfoList().
The tests for it has been moved testDBTablesHosts.
