+++ server/src/DBTablesMonitoring.h
@@ -142,10 +142,6 @@ public:
 	DBTablesMonitoring(DBAgent &dbAgent);
 	virtual ~DBTablesMonitoring();
 
-	void getHostInfoList(HostInfoList &hostInfoList,
-	                     const HostsQueryOption &option)
-	  __attribute__ ((deprecated));
-
 	void addTriggerInfo(TriggerInfo *triggerInfo);
 	void addTriggerInfoList(const TriggerInfoList &triggerInfoList);
