+++ server/test/testDBTablesHost.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2014 Project Hatohol
+ * Copyright (C) 2014-2015 Project Hatohol
  *
  * This file is part of Hatohol.
  *
@@ -24,6 +24,7 @@
 #include "DBTablesHost.h"
 #include "Hatohol.h"
 #include "Helpers.h"
+#include "testDBTablesMonitoring.h"
 using namespace std;
 using namespace mlpl;
 
@@ -34,6 +35,46 @@ namespace testDBTablesHost {
 	DBTablesHost &VAR_NAME = _dbHatohol.getDBTablesHost();
 
 static const char *testHostName = "FOO FOO FOO";
+
+struct AssertGetHostsArg
+  : public AssertGetHostResourceArg<ServerHostDef, HostsQueryOption>
+{
+	HostInfoList expectedHostList;
+
+	AssertGetHostsArg(gconstpointer ddtParam)
+	{
+		fixtures = testServerHostDef;
+		numberOfFixtures = NumTestServerHostDef;
+		setDataDrivenTestParam(ddtParam);
+	}
+
+	// This should be moved to AssertGetHostResourceArg later
+	virtual bool isAuthorized(const ServerHostDef &svHostDef) override
+	{
+		return ::isAuthorized(userId, svHostDef.hostId);
+	}
+
+	virtual HostIdType getHostId(const ServerHostDef &svHostDef) const override
+	{
+		HostIdType hostId;
+		cppcut_assert_equal(
+		  1, sscanf(svHostDef.hostIdInServer.c_str(), "%" FMT_HOST_ID,
+		            &hostId));
+		return hostId;
+	}
+
+	virtual string makeOutputText(const ServerHostDef &svHostDef)
+	{
+		string expectedOut =
+		  StringUtils::sprintf(
+		    "%" PRIu32 "|%s|%s\n",
+		    svHostDef.serverId, svHostDef.hostIdInServer.c_str(),
+		    svHostDef.name.c_str());
+		return expectedOut;
+	}
+
+};
+
 
 static void assertUpsertHost(const HostIdType &hostId)
 {
@@ -65,7 +106,24 @@ static void assertUpsertHost(const HostIdType &hostId)
 	assertDBContent(&dbHost.getDBAgent(), statement, expect);
 }
 
+static void _assertGetHosts(AssertGetHostsArg &arg)
+{
+	loadTestDBUser();
+	loadTestDBAccessList();
+	loadTestDBServerHostDef();
+	loadTestDBHostgroupMember();
 
+	DECLARE_DBTABLES_HOST(dbHost);
+	arg.fixup();
+	ServerHostDefVect svHostDefs;
+	dbHost.getServerHostDefs(svHostDefs, arg.option);
+
+	ServerHostDefVectConstIterator svHostItr = svHostDefs.begin();
+	for (; svHostItr != svHostDefs.end(); ++svHostItr)
+		arg.actualRecordList.push_back(*svHostItr);
+	arg.assert();
+}
+#define assertGetHosts(A) cut_trace(_assertGetHosts(A))
 
 void cut_setup(void)
 {
@@ -723,6 +781,65 @@ void test_getServerHostDefs(gconstpointer data)
 		cppcut_assert_equal(exp.name, act.name);
 	}
 	cppcut_assert_equal(true, expectIds.empty());
+}
+
+void data_getHostInfoList(void)
+{
+	prepareTestDataForFilterForDataOfDefunctServers();
+}
+
+void test_getHostInfoList(gconstpointer data)
+{
+	AssertGetHostsArg arg(data);
+	assertGetHosts(arg);
+}
+
+void data_getHostInfoListForOneServer(void)
+{
+	prepareTestDataForFilterForDataOfDefunctServers();
+}
+
+void test_getHostInfoListForOneServer(gconstpointer data)
+{
+	AssertGetHostsArg arg(data);
+	arg.targetServerId = 1;
+	assertGetHosts(arg);
+}
+
+void data_getHostInfoListWithNoAuthorizedServer(void)
+{
+	prepareTestDataForFilterForDataOfDefunctServers();
+}
+
+void test_getHostInfoListWithNoAuthorizedServer(gconstpointer data)
+{
+	AssertGetHostsArg arg(data);
+	arg.userId = 4;
+	assertGetHosts(arg);
+}
+
+void data_getHostInfoListWithOneAuthorizedServer(gconstpointer data)
+{
+	prepareTestDataForFilterForDataOfDefunctServers();
+}
+
+void test_getHostInfoListWithOneAuthorizedServer(gconstpointer data)
+{
+	AssertGetHostsArg arg(data);
+	arg.userId = 5;
+	assertGetHosts(arg);
+}
+
+void data_getHostInfoListWithUserWhoCanAccessSomeHostgroups(void)
+{
+	prepareTestDataForFilterForDataOfDefunctServers();
+}
+
+void test_getHostInfoListWithUserWhoCanAccessSomeHostgroups(gpointer data)
+{
+	AssertGetHostsArg arg(data);
+	arg.userId = 3;
+	assertGetHosts(arg);
 }
 
 } // namespace testDBTablesHost
